Extracts InvalidError into a separate file and makes it a subclass of Ember.Error
